+++ src/tripCard/TripCard.story.mdx
@@ -1,0 +1,93 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
+
+import { WarningIcon } from '../icon/warningIcon'
+import { TripCard } from '.'
+
+<Meta title="Widgets|TripCard" />
+
+# **TripCard**
+
+<Preview>
+  <Story name="Default">
+    <TripCard
+      aria-label="Pick-up point: Paris, Drop-off point: Bordeaux, Departure time: 09:00, warning"
+      href="/"
+      itinerary={[
+        {
+          mainLabel:  'Paris',
+          subLabel:  'Porte de Vincennes',
+          time:  '09:00',
+          isoDate: '2017-12-11T09:00',
+          distanceFromPoint: '1,5km',
+        },
+        {
+          mainLabel:  'Bordeaux',
+          subLabel: 'Gare Bordeaux Saint-Jean',
+          time:  '12:00',
+          isoDate: '2017-12-11T12:00',
+          distanceFromPoint: '8km',
+        },
+      ]}
+      price="8,00€"
+      flags={{
+        ladiesOnly: true,
+        aloneInTheBack: true,
+        maxTwo: true,
+        autoApproval: true,
+      }}
+      titles={{
+        ladiesOnly: 'Ladies Only',
+        aloneInTheBack: 'Alone in the back',
+        maxTwo: 'Max 2 in the back',
+        autoApproval: 'Auto approval',
+      }}
+      metaUrl="Meta URL"
+      statusInformation={{
+        icon: <WarningIcon />,
+        text: 'Warning',
+        highlighted: false,
+      }}
+      badge="The Closest"
+      title="Sun march 8th, 18:00"
+      driver={{
+          avatarUrl: '//placehold.it/500x500',
+            firstName: 'Jane',
+            rating:  '4.9',
+        }} />
+  </Story>
+</Preview>
+
+## Specifications
+
+> Card
+
+This component serves as a container for other components. Its entire area becomes clickable.
+
+Components inside cards have a medium 16 pt inner margin as compared to regular state - 24 pt inner margin.
+
+> Shadow
+
+To further establish unique character of the card and it’s content - this component has a material like appearance.
+
+The card casts shadow:
+`box-shadow: 0 1pt 4pt rgba(0,0,0,0.16), 0 2pt 8pt rgba(0,0,0,0.08);`
+
+Additionaly on desktop when it receives a cursor :hover
+`box-shadow: 0 2pt 8pt rgba(0,0,0,0.08), 0 2pt 16pt rgba(0,0,0,0.08);`
+
+## When should I use it?
+
+Use cards to represent a summary of a single trip in a collection view.
+Use backgroundless components to represent card’s content.
+
+- ✅ Always clickable.
+
+## Usage
+
+```jsx
+import { TripCard } from '@blablacar/ui-library/build/tripCard'
+
+<TripCard />
+```
+
+<Props of={TripCard} />
